feat: add unstaking notice banner with warning styles
