+++ public/locales/en/staking.json
@@ -53,5 +53,7 @@
   "approveButton": "Approve {{symbol}}",
   "approvePending": "Approving {{amount}} {{symbol}}...",
   "approveSuccess": "Successfully approved {{amount}} {{symbol}}",
-  "approveError": "Failed to approve tokens"
+  "approveError": "Failed to approve tokens",
+  "unstakingInformationTitle": "Unstaking Notice",
+  "unstakingInformationDescription": "Unstaking will forfeit any pending rewards for the \n current period."
 }
